+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+- Explicit confirmation is now required to prevent accidental destroying
+  of the cluster with `pcs cluster destroy` ([rhbz#1283805])
+
 ### Changed
 - Pcs no longer depends on python3-distro package
 - 'pcs status xml' now prints cluster status in the new format provided by
@@ -14,6 +18,7 @@
 - Deprecated obsolete commands `pcs config import-cman` and `pcs config export
   pcs-commands|pcs-commands-verbose` have been removed ([rhbz#1881064])
 
+[rhbz#1283805]: https://bugzilla.redhat.com/show_bug.cgi?id=1283805
 [rhbz#1881064]: https://bugzilla.redhat.com/show_bug.cgi?id=1881064
 [rhbz#1985981]: https://bugzilla.redhat.com/show_bug.cgi?id=1985981